move migrations to separate file
