+++ internal/store/postgres_migrations.go
@@ -1,0 +1,35 @@
+package store
+
+import (
+	"context"
+	"fmt"
+)
+
+func (p *PostgresStore) createSchema(ctx context.Context) error {
+	tables := []string{
+		"CREATE EXTENSION IF NOT EXISTS pgcrypto",
+		"CREATE TABLE IF NOT EXISTS users (" +
+			"id SERIAL PRIMARY KEY, " +
+			"username TEXT NOT NULL UNIQUE, " +
+			"password TEXT NOT NULL)",
+		"CREATE TABLE IF NOT EXISTS secrets (" +
+			"id SERIAL PRIMARY KEY, " +
+			"user_id INTEGER NOT NULL, " +
+			"data BYTEA NOT NULL, " +
+			"kind INTEGER NOT NULL, " +
+			"FOREIGN KEY (user_id) REFERENCES users (id) ON DELETE CASCADE)",
+		"CREATE TABLE IF NOT EXISTS meta (" +
+			"id SERIAL PRIMARY KEY, " +
+			"secret_id INTEGER NOT NULL, " +
+			"text TEXT NOT NULL, " +
+			"FOREIGN KEY (secret_id) REFERENCES secrets (id) ON DELETE CASCADE)",
+	}
+	for _, table := range tables {
+		_, err := p.database.ExecContext(ctx, table)
+		if err != nil {
+			return fmt.Errorf("failed to execute statement: %w", err)
+		}
+	}
+
+	return nil
+}
